feat(auto-populate): enhance on-chain data auto-population features
- Updated configuration options in `env.example` and `config.py` to allow for flexible on-chain symbol and metrics population.
- Implemented fetching of top market coins from CoinGecko when no symbols are configured.
- Adjusted database schema to support new data types for on-chain metrics.
- Enhanced the auto-population process to include detailed logging and progress tracking.
- Added functionality to ensure trending assets are automatically managed in the frontend.

+++ backend/app/db/models/on_chain_metrics.py
@@ -3,7 +3,7 @@
 import uuid
 from datetime import datetime
 
-from sqlalchemy import DateTime, ForeignKey, Numeric, String, Uuid
+from sqlalchemy import DateTime, Double, ForeignKey, String, Uuid
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.base import Base
@@ -21,4 +21,4 @@ class OnChainMetric(Base):
         primary_key=True,
     )
     metric_name: Mapped[str] = mapped_column(String(50), primary_key=True)
-    value: Mapped[float] = mapped_column(Numeric(24, 8), nullable=False)
+    value: Mapped[float] = mapped_column(Double, nullable=False)
